requirements.txt
flask==2.0.1
python-dotenv==0.19.2
groq==0.4.0
